requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paho-mqtt==1.6.1
pyudev==0.24.1
httpx==0.26.0
click==8.1.7
pyyaml==6.0.1
jinja2==3.1.3
aiosqlite==0.19.0
python-multipart==0.0.6
smbprotocol==1.12.0
watchdog==3.0.0
